feat : 주식 수 300개로 늘리고, Docker 및 모니터링 환경 구축

+++ src/main/java/com/example/stocksimulation/domain/entity/Stock.java
@@ -19,7 +19,7 @@ public class Stock {
     }
 
     @Builder
-    public Stock(String code, int price, String name) {
+    public Stock(String code, long price, String name) {
         this.code= code;
         this.price = price;
         this.name = name;
